+++ src/LiveComponent/tests/Unit/Metadata/LiveComponentMetadataTest.php
@@ -15,6 +15,7 @@ use PHPUnit\Framework\TestCase;
 use Symfony\UX\LiveComponent\Attribute\LiveProp;
 use Symfony\UX\LiveComponent\Metadata\LiveComponentMetadata;
 use Symfony\UX\LiveComponent\Metadata\LivePropMetadata;
+use Symfony\UX\LiveComponent\Metadata\UrlMapping;
 use Symfony\UX\TwigComponent\ComponentMetadata;
 
 class LiveComponentMetadataTest extends TestCase
@@ -36,5 +37,20 @@ class LiveComponentMetadataTest extends TestCase
         $expected = ['yesUpdateFromParent1' => 'bar'];
         $actual = $liveComponentMetadata->getOnlyPropsThatAcceptUpdatesFromParent($inputProps);
         $this->assertEquals($expected, $actual);
+    }
+
+    public function testGetAllUrlMappings(): void
+    {
+        $aliasUrlMapping = new UrlMapping('alias');
+        $propMetadas = [
+            new LivePropMetadata('noUrlMapping', new LiveProp(), null, false, false, null),
+            new LivePropMetadata('basicUrlMapping', new LiveProp(url: true), null, false, false, null),
+            new LivePropMetadata('aliasUrlMapping', new LiveProp(url: $aliasUrlMapping), null, false, false, null),
+        ];
+        $liveComponentMetadata = new LiveComponentMetadata(new ComponentMetadata([]), $propMetadas);
+        $urlMappings = $liveComponentMetadata->getAllUrlMappings();
+        $this->assertCount(2, $urlMappings);
+        $this->assertInstanceOf(UrlMapping::class, $urlMappings['basicUrlMapping']);
+        $this->assertEquals($aliasUrlMapping, $urlMappings['aliasUrlMapping']);
     }
 }